Move ScrapingUtilities to FanGraphs.utilities
