+++ FanGraphs/selectors/__init__.py
@@ -1,6 +1,2 @@
 #! python3
 # FanGraphs/selectors/__init__.py
-
-"""
-Contains packages of modules of CSS selectors.
-"""
