+++ FanGraphs/tests/__init__.py
@@ -1,6 +1,2 @@
 #! python3
 # tests/__init__.py
-
-"""
-Modules of tests for the modules in :py:mod:`FanGraphs`.
-"""
