--- docs/selectors/leaders_sel.rst
@@ -1,7 +1,0 @@
-FanGraphs.selectors.leaders_sel
-==============================================================================
-
-.. toctree::
-
-.. automodule:: FanGraphs.selectors.leaders_sel
-    :members:
