+++ docs/utilities.rst
@@ -1,0 +1,7 @@
+FanGraphs.utilities
+==============================================================================
+
+.. automodule:: FanGraphs.utilities
+    :members:
+    :undoc-members:
+    :show-inheritance:
